+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased][unreleased]
+
+
+## [0.1.3] - 2016-08-05
 ### Added
 - Covered code with tests
 
@@ -27,6 +30,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - ``wscheck`` lib with command line bin
 
 
-[unreleased]: https://github.com/andras-tim/wscheck/compare/v0.1.2...HEAD
+[unreleased]: https://github.com/andras-tim/wscheck/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/andras-tim/wscheck/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/andras-tim/wscheck/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/andras-tim/wscheck/compare/v0.1.0...v0.1.1